requirements.txt
flask>=3.0.0
google-genai>=0.5.0
python-dotenv>=1.0.0
